Sketched 'BigBedFile.write'
  tests SHOULD fail

+++ src/test/kotlin/org/jbb/big/BigFileTest.kt
@@ -8,7 +8,7 @@ public class BigFileTest {
         BigBedFile.read(Examples.get("example1.bb")).use { bbf ->
             val header = bbf.header
             assertEquals(1.toShort(), header.version)
-            assertEquals(5, header.zoomLevelCount)
+            assertEquals(5.toShort(), header.zoomLevelCount)
             assertEquals(3.toShort(), header.fieldCount)
             assertEquals(3.toShort(), header.definedFieldCount)
             assertEquals(0, header.uncompressBufSize)
